+++ svcbatch.c
@@ -983,20 +983,24 @@ static DWORD openlogfile(BOOL firstopen)
 #endif
     memset(sfx, 0, 48);
     if (GetFileAttributesExW(logfilename, GetFileExInfoStandard, &ad)) {
-        DWORD mm = MOVEFILE_REPLACE_EXISTING;
+        DWORD mm, ac = 0;
 
+retry:
         if (firstopen)
             reportsvcstatus(SERVICE_START_PENDING, SVCBATCH_START_HINT);
         if (autorotate) {
             SYSTEMTIME st;
-
-            FileTimeToSystemTime(&ad.ftLastWriteTime, &st);
+            if (ac)
+                GetSystemTime(&st);
+            else
+                FileTimeToSystemTime(&ad.ftLastWriteTime, &st);
             _snwprintf(sfx, 20, L".%.4d-%.2d-%.2d.%.2d%.2d%.2d",
                        st.wYear, st.wMonth, st.wDay,
                        st.wHour, st.wMinute, st.wSecond);
             mm = 0;
         }
         else {
+            mm = MOVEFILE_REPLACE_EXISTING;
             sfx[0] = L'.';
             sfx[1] = L'0';
         }
@@ -1004,7 +1008,15 @@ static DWORD openlogfile(BOOL firstopen)
         if (!MoveFileExW(logfilename, logpb, mm)) {
             rc = GetLastError();
             xfree(logpb);
-            return svcsyserror(__LINE__, rc, L"MoveFileExW");
+            if (autorotate && ac++ == 0) {
+#if defined(_DBGVIEW)
+                dbgprintf(__FUNCTION__, "Retrying %S%S rc=%lu", logfilename, sfx, rc);
+#endif
+                goto retry;
+            }
+            else {
+                return svcsyserror(__LINE__, rc, L"MoveFileExW");
+            }
         }
     }
     else {