+++ svcbatch.h
@@ -36,7 +36,7 @@
  */
 #define SVCBATCH_MAJOR_VERSION  1
 #define SVCBATCH_MINOR_VERSION  0
-#define SVCBATCH_PATCH_VERSION  6
+#define SVCBATCH_PATCH_VERSION  7
 #if defined(_VENDOR_NUM)
 # define SVCBATCH_MICRO_VERSION _VENDOR_NUM
 #else
@@ -46,7 +46,7 @@
 /**
  * Set to zero for non dev versions
  */
-#define SVCBATCH_ISDEV_VERSION  0
+#define SVCBATCH_ISDEV_VERSION  1
 
 #define SVCBATCH_NAME           "SvcBatch"
 #define SVCBATCH_SVCNAME        "SvcBatch Service"