+++ CHANGES.md
@@ -6,6 +6,10 @@ For a full list of changes, see the [git commit log][log]
   [log]: https://github.com/mturk/svcbatch/commits/
 
 
+## v1.0.7
+
+ * In development
+
 ## v1.0.6
 
  * Fix log rotation timeout processing